aula-4/client/requirements.txt
docopt==0.6.2
fastavro==1.4.3
kafka-python==2.0.2
python-dotenv==0.18.0
